chore: Add Windows to Travis
This commit runs the unit tests on Windows. There is one test that is
failing on Windows that we skip. I spoke to Mathias B and we agreed to
defer debugging this for now in favour of getting tests running on
Windows. But we didn't want to ignore it forever, hence giving the test
a date at which it will start to fail.